+++ genericsuite/lib/services/app_callables_super.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+
+// import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 import 'theme_config_defaults.dart';
 
@@ -35,7 +36,7 @@ class AppCallablesSuper {
   }
 
   Map<String, dynamic> getUserCallbacks(
-    FlutterSecureStorage storage,
+    // FlutterSecureStorage storage,
     BuildContext context,
   ) {
     return {'specificFunctions': {}, "components": {}, "childComponents": {}};