+++ genericsuite/lib/services/autocomplete_service.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 import 'http_service.dart';
 import 'message_service.dart';
@@ -12,14 +11,14 @@ const acsDebug = false;
 const Duration debounceDuration = Duration(milliseconds: 500);
 
 class AsyncAutocomplete extends StatefulWidget {
-  final FlutterSecureStorage storage;
+  // final FlutterSecureStorage storage;
   final Map<String, dynamic> config;
   final String value;
   final Function(String) onSelected;
   final Function(String, String, int) setError;
 
   const AsyncAutocomplete({
-    required this.storage,
+    // required this.storage,
     required this.config,
     required this.value,
     required this.onSelected,
@@ -53,7 +52,7 @@ class AsyncAutocompleteState extends State<AsyncAutocomplete> {
     Iterable<Map<String, dynamic>> optionsRetrieved = {};
     try {
       optionsRetrieved = await _ApiCall(
-        widget.storage,
+        // widget.storage,
         widget.config,
         widget.setError,
       ).search(_currentQuery!);
@@ -117,11 +116,12 @@ class AsyncAutocompleteState extends State<AsyncAutocomplete> {
 
 // Remote API
 class _ApiCall {
-  FlutterSecureStorage storage;
+  // FlutterSecureStorage storage;
   Map<String, dynamic> widgetConfig;
   Function(String, String, int) setError;
 
-  _ApiCall(this.storage, this.widgetConfig, this.setError);
+  // _ApiCall(this.storage, this.widgetConfig, this.setError);
+  _ApiCall(this.widgetConfig, this.setError);
 
   // Searches the options, but injects a fake "network" delay.
   Future<Iterable<Map<String, dynamic>>> search(String query) async {
@@ -195,7 +195,8 @@ class _ApiCall {
     dynamic bodyParams,
     Map<String, dynamic> getParams = const {},
   }) async {
-    HttpUtilities api = HttpUtilities(storage);
+    // HttpUtilities api = HttpUtilities(storage);
+    HttpUtilities api = HttpUtilities();
     Map<String, dynamic> body = bodyParams.cast<String, dynamic>();
     if (acsDebug) {
       logDebug(
@@ -217,14 +218,14 @@ class _ApiCall {
 }
 
 class SuggestionDropdown extends StatefulWidget {
-  final FlutterSecureStorage storage;
+  // final FlutterSecureStorage storage;
   final Map<String, dynamic> config;
   final String value;
   final Function(String) onSelected;
   final Function(String, String, int) setError;
 
   const SuggestionDropdown({
-    required this.storage,
+    // required this.storage,
     required this.config,
     required this.value,
     required this.onSelected,
@@ -284,7 +285,7 @@ class SuggestionDropdownState extends State<SuggestionDropdown> {
       children: <Widget>[
         Text(widget.config['label']),
         AsyncAutocomplete(
-          storage: widget.storage,
+          // storage: widget.storage,
           config: widget.config,
           value: widget.value,
           onSelected: widget.onSelected,