+++ genericsuite/lib/services/form_field_service.dart
@@ -2,7 +2,8 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+
+// import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 import 'autocomplete_service.dart';
 import 'select_options_service.dart';
@@ -157,7 +158,7 @@ class _PasswordFieldState extends State<PasswordField> {
 }
 
 class DataFormBody extends StatefulWidget {
-  final FlutterSecureStorage storage;
+  // final FlutterSecureStorage storage;
   final Map<String, dynamic> editorConfig;
   final Map<String, dynamic> constants;
   final Map<String, dynamic> selectedItem;
@@ -171,7 +172,7 @@ class DataFormBody extends StatefulWidget {
 
   const DataFormBody({
     Key? key,
-    required this.storage,
+    // required this.storage,
     required this.editorConfig,
     required this.constants,
     required this.selectedItem,
@@ -573,7 +574,7 @@ class _DataFormBodyState extends State<DataFormBody> {
         case 'suggestion_dropdown':
           formFields.add(
             SuggestionDropdown(
-              storage: widget.storage,
+              // storage: widget.storage,
               config: fieldElement,
               value: fieldElementValue,
               onSelected: (value) {