+++ genericsuite/lib/services/redirect_service.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+
+// import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 import '../views/homepage.dart';
 import 'app_callables_super.dart';
@@ -9,7 +10,7 @@ const redirectToCrud = false;
 const rdSvDebug = false;
 
 Widget getMainScreen(
-  storage,
+  // storage,
   HomePageBodyBuilder homePageBodyBuilder,
   AlternateWidgetBuilder alternateWidgetBuilder,
   AppCallablesSuper appCallables,
@@ -18,13 +19,14 @@ Widget getMainScreen(
     if (rdSvDebug) {
       logDebug('>>>> Redirect to CRUD...');
     }
-    return alternateWidgetBuilder(storage);
+    // return alternateWidgetBuilder(storage);
+    return alternateWidgetBuilder();
   } else {
     if (rdSvDebug) {
       logDebug('>>>> Redirect to HomePage...');
     }
     return HomePage(
-      storage,
+      // storage,
       homePageBodyBuilder,
       alternateWidgetBuilder,
       appCallables,
@@ -33,7 +35,7 @@ Widget getMainScreen(
 }
 
 void redirectMainScreen(
-  FlutterSecureStorage storage,
+  // FlutterSecureStorage storage,
   BuildContext context,
   HomePageBodyBuilder homePageBodyBuilder,
   AlternateWidgetBuilder alternateBodyBuilder,
@@ -43,7 +45,7 @@ void redirectMainScreen(
     context,
     MaterialPageRoute(
       builder: (context) => getMainScreen(
-        storage,
+        // storage,
         homePageBodyBuilder,
         alternateBodyBuilder,
         appCallables,