+++ genericsuite/lib/views/homepage.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+
+// import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 import '../services/app_callables_super.dart';
 import '../services/current_user_service.dart';
@@ -10,21 +11,28 @@ import '../widgets/app_frame.dart';
 
 const debug = false;
 
-typedef HomePageBodyBuilder = Widget Function(
-    FlutterSecureStorage storage, Map<String, dynamic> userData);
+typedef HomePageBodyBuilder =
+    Widget Function(
+      // FlutterSecureStorage storage,
+      Map<String, dynamic> userData,
+    );
 
-typedef AlternateWidgetBuilder = Widget Function(FlutterSecureStorage storage);
+// typedef AlternateWidgetBuilder = Widget Function(FlutterSecureStorage storage);
+typedef AlternateWidgetBuilder = Widget Function();
 
 class HomePage extends StatefulWidget {
-  final FlutterSecureStorage storage;
+  // final FlutterSecureStorage storage;
   final HomePageBodyBuilder homePageBodyBuilder;
   final AlternateWidgetBuilder alternateWidgetBuilder;
   final AppCallablesSuper appCallables;
 
-  const HomePage(this.storage, this.homePageBodyBuilder,
-      this.alternateWidgetBuilder, this.appCallables,
-      {Key? key})
-      : super(key: key);
+  // const HomePage(this.storage, this.homePageBodyBuilder,
+  const HomePage(
+    this.homePageBodyBuilder,
+    this.alternateWidgetBuilder,
+    this.appCallables, {
+    super.key,
+  });
 
   @override
   State<HomePage> createState() => _HomePageState();
@@ -58,10 +66,13 @@ class _HomePageState extends State<HomePage> {
   }
 
   void showScaffoldMessages(
-      BuildContext context, AppCallablesSuper appCallables) {
+    BuildContext context,
+    AppCallablesSuper appCallables,
+  ) {
     if (debug) {
       logDebug(
-          'HomePage | showScaffoldMessages | errorMessage: $errorMessage | infoMessage: $infoMessage');
+        'HomePage | showScaffoldMessages | errorMessage: $errorMessage | infoMessage: $infoMessage',
+      );
     }
     if (errorMessage.isNotEmpty) {
       showScaffoldMessage(errorMessage, context, typeError, appCallables);
@@ -83,35 +94,40 @@ class _HomePageState extends State<HomePage> {
       // Cache user data to avoid repeated API calls to "/users/current_user_d"
       return Future.value(userData);
     }
-    return getCurrentUserData(widget.storage).then((data) {
-      if (data['errorMessage'].isNotEmpty) {
-        errorMessage = data['errorMessage'];
-        return Future.value(userData);
-      }
-      userData = data['userData'];
-      return Future.value(userData);
-    }).catchError((error) {
-      errorMessage = error.toString();
-      return Future.value(userData);
-    });
+    // return getCurrentUserData(widget.storage).then((data) {
+    return getCurrentUserData()
+        .then((data) {
+          if (data['errorMessage'].isNotEmpty) {
+            errorMessage = data['errorMessage'];
+            return Future.value(userData);
+          }
+          userData = data['userData'];
+          return Future.value(userData);
+        })
+        .catchError((error) {
+          errorMessage = error.toString();
+          return Future.value(userData);
+        });
   }
 
   Widget buildHomePage(BuildContext context) {
     // const title = "Dashboard";
     return AppFrame(
-      storage: widget.storage,
+      // storage: widget.storage,
       appCallables: widget.appCallables,
       body: Center(
         child: FutureBuilder(
-            future: loadHomeData(true),
-            builder: (context, snapshot) =>
-                snapshot.hasData && errorMessage.isEmpty
-                    ? widget.homePageBodyBuilder(widget.storage, userData)
-                    : snapshot.hasError || errorMessage.isNotEmpty
-                        ? snapshot.hasError
-                            ? Text(setErrorMessage(snapshot.error.toString()))
-                            : Text(setErrorMessage("Error loading data"))
-                        : const Center(child: CircularProgressIndicator())),
+          future: loadHomeData(true),
+          builder: (context, snapshot) =>
+              snapshot.hasData && errorMessage.isEmpty
+              // ? widget.homePageBodyBuilder(widget.storage, userData)
+              ? widget.homePageBodyBuilder(userData)
+              : snapshot.hasError || errorMessage.isNotEmpty
+              ? snapshot.hasError
+                    ? Text(setErrorMessage(snapshot.error.toString()))
+                    : Text(setErrorMessage("Error loading data"))
+              : const Center(child: CircularProgressIndicator()),
+        ),
       ),
     );
   }