+++ genericsuite/lib/widgets/app_drawer.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+// import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:genericsuite/services/app_callables_super.dart';
 
 import '../services/http_service.dart';
@@ -15,11 +15,11 @@ const acceptedUserGroups = ['users'];
 const defaultUserGroup = 'users';
 
 class AppDrawer extends StatefulWidget {
-  final FlutterSecureStorage storage;
+  // final FlutterSecureStorage storage;
   final AppCallablesSuper appCallables;
 
-  const AppDrawer({Key? key, required this.storage, required this.appCallables})
-    : super(key: key);
+  // const AppDrawer({Key? key, required this.storage, required this.appCallables})
+  const AppDrawer({super.key, required this.appCallables});
 
   @override
   State<AppDrawer> createState() => _AppDrawerState();
@@ -119,7 +119,8 @@ class _AppDrawerState extends State<AppDrawer> {
     setState(() {
       _isLoading = true;
     });
-    return loadConfig(widget.storage).then((configStr) {
+    // return loadConfig(widget.storage).then((configStr) {
+    return loadConfig().then((configStr) {
       return getJsonFileList(configFilename).then((menuConfigRaw) {
         Map<String, dynamic> config = json.decode(configStr);
         configItems = config["configItems"];
@@ -292,13 +293,15 @@ class _AppDrawerState extends State<AppDrawer> {
                     item['callable']['function'] != null)) {
               Navigator.pop(context); // Close the drawer
               if (item['callable']['function'] != null) {
-                item['callable']['function'](context, widget.storage);
+                // item['callable']['function'](context, widget.storage);
+                item['callable']['function'](context);
               } else {
                 Navigator.push(
                   context,
                   MaterialPageRoute(
                     builder: (context) =>
-                        item['callable']['widget'](widget.storage),
+                        // item['callable']['widget'](widget.storage),
+                        item['callable']['widget'](),
                   ),
                 );
               }