+++ genericsuite/lib/widgets/error_reporter_widget.dart
@@ -7,7 +7,8 @@ const showScaffold = true;
 class ErrorReporter extends StatelessWidget {
   final String message;
 
-  const ErrorReporter({Key? key, required this.message}) : super(key: key);
+  // const ErrorReporter({Key? key, required this.message}) : super(key: key);
+  const ErrorReporter({super.key, required this.message});
 
   @override
   Widget build(BuildContext context) {